Final Push
Ready for review, the entire page now has a background image that does not scroll, whilst remaining horizontally and vertically centered. The body and body header both have set background colors. The asides have rounded corners, and the headings within them are rounded on the top. The asides also have perspective so that they appear to go to a center. A 'cube' was created, where with different positioning, uniform heights and widths, and various transforms, the images appear to be in a cube, where they still are facing the user.